+++ CHANGELOG.md
@@ -6,10 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `start.wav`
+- `well_done.wav`
+- `preparing.wav`
 - `requirements-splitter.py`
 - 4-7-8 program
 - 7-11 program
 ### Changed
+- Sound speaker
 - Test system modified
 - Menu optimized
 ## [0.3] - 2021-02-09